next/feature/guestbook
Separated Receipt into ReceiptInputs and ReceiptDisplay.
Added DateInput, ImageInput and ServiceInput for Receipt page.
