+++ src/components/Shared/Inputs/ServiceInput.tsx
@@ -1,0 +1,90 @@
+import { ChangeEvent, useCallback } from "react";
+import { useTranslation } from "react-i18next";
+
+type Props = {
+  key?: any;
+  serviceTotal: string;
+  setServiceName: (serviceName: string) => void;
+  setServiceAmount: (serviceAmount: string) => void;
+  setServicePrice: (servicePrice: string) => void;
+  removeService: () => void;
+};
+
+const ServiceInput = (props: Props) => {
+  const {
+    key,
+    setServiceName,
+    setServiceAmount,
+    setServicePrice,
+    removeService,
+    serviceTotal,
+  } = props;
+  const { t, i18n } = useTranslation("ServiceInput");
+
+  const handleServiceNameChange = useCallback(
+    (e: ChangeEvent<HTMLInputElement>) => {
+      setServiceName(e.target.value);
+    },
+    [setServiceName]
+  );
+
+  const handleServiceAmountChange = useCallback(
+    (e: ChangeEvent<HTMLInputElement>) => {
+      setServiceAmount(e.target.value);
+    },
+    [setServiceAmount]
+  );
+
+  const handleServicePriceChange = useCallback(
+    (e: ChangeEvent<HTMLInputElement>) => {
+      setServicePrice(e.target.value);
+    },
+    [setServicePrice]
+  );
+
+  const handleRemoveService = useCallback(() => {
+    removeService();
+  }, [removeService]);
+
+  return (
+    <div key={`service-${key}`} className="flex flex-row gap-2 w-full mb-2">
+      <input
+        className="appearance-none border rounded-md text-gray-700 leading-tight focus:border-blue-500 w-2/6"
+        type={"text"}
+        placeholder={t("name")}
+        onChange={e => setServiceName(e.target.value)}
+      />
+      <input
+        className="appearance-none border rounded-md text-gray-700 leading-tight focus:border-blue-500 w-1/6"
+        type={"number"}
+        lang={i18n.language}
+        placeholder={t("amount")}
+        onChange={e => setServiceAmount(e.target.value)}
+      />
+      <input
+        className="appearance-none border rounded-md text-gray-700 leading-tight focus:border-blue-500 w-1/6"
+        type={"number"}
+        lang={i18n.language}
+        placeholder={t("price")}
+        onChange={e => setServicePrice(e.target.value)}
+      />
+      <div className="appearance-none bg-gray-200 flex items-center justify-center border rounded-md text-gray-700 leading-tight focus:border-blue-500 w-1/6">
+        {Number(serviceTotal).toLocaleString(i18n.language, {
+          minimumFractionDigits: 2,
+        })}
+      </div>
+      <button
+        className={`w-10 h-10`}
+        style={{
+          backgroundImage: "url(/Styles/Assets/Images/xCircle.svg)",
+          backgroundRepeat: "no-repeat",
+          backgroundPosition: "center",
+          backgroundSize: "contain",
+        }}
+        onClick={handleRemoveService}
+      />
+    </div>
+  );
+};
+
+export default ServiceInput;
