+++ src/components/Shared/Inputs/i18n/ServiceInput.en-US.json
@@ -1,0 +1,5 @@
+{
+  "name": "Name",
+  "amount": "Amount",
+  "price": "Price"
+}
