+++ src/components/Shared/Inputs/i18n/ServiceInput.hr.json
@@ -1,0 +1,5 @@
+{
+  "name": "Naziv",
+  "amount": "Količina",
+  "price": "Cijena"
+}
