+++ src/components/Shared/Receipt/ReceiptDisplay/ReceiptDisplay.module.scss
@@ -1,0 +1,67 @@
+.document-controls {
+  opacity: 0;
+  bottom: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  transition: opacity ease-in-out 0.2s;
+  position: absolute;
+  display: flex;
+  place-content: center;
+  align-items: flex-end;
+  height: 100%;
+  width: 100%;
+  padding: 20px;
+}
+.page-controls-navigation {
+  border-radius: 4px;
+  background-color: #ffffff;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.3);
+  text-align: center;
+  width: fit-content;
+  height: fit-content;
+}
+.page-controls-navigation span {
+  font-size: 14px;
+  font-weight: bold;
+  color: rgb(0, 0, 0);
+  margin: 0px 10px;
+}
+.page-controls-navigation button {
+  width: 40px;
+  height: 40px;
+}
+.page-controls-navigation button:hover {
+  background-color: #e6e6e6;
+  cursor: pointer;
+  border-radius: 4px;
+}
+
+.documentPDFView {
+  text-align: -webkit-center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.pdf-download {
+  background: url("/Styles/Assets/Images/downloadDocument.svg");
+  background-repeat: no-repeat;
+  background-position: center;
+  background-color: #ffffff;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.3);
+  width: 40px;
+  height: 40px;
+  border-radius: 4px;
+  cursor: pointer;
+  margin-left: 10px;
+}
+
+.pdf-download:hover {
+  background-color: #e6e6e6;
+}
+
+.documentPDFView:hover .document-controls,
+.document-controls:hover {
+  opacity: 1;
+}
