+++ src/components/Shared/Receipt/ReceiptInputs/ReceiptInputs.module.scss
@@ -11,10 +11,3 @@
 .section {
   height: 100%;
 }
-
-.removeButton {
-  background-image: url('/Styles/Assets/Images/xCircle.svg');
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: contain;
-}
