+++ src/components/Shared/Receipt/ReceiptInputs/i18n/ReceiptInputs.en-US.json
