+++ src/components/Shared/Receipt/Templates/ReceiptTemplate.tsx
@@ -1,38 +1,12 @@
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
+import { TransactionReceiptData } from "../Receipt";
 import TransactionReceipt from "./TransactionReceipt";
 
 type Props = {
-  apartmentData: {
-    name: string;
-    address: string;
-    owner: string;
-    image: string;
-    pid: string;
-    iban: string;
-  };
-  recepientData: {
-    recepientName: string;
-    recepientAddress: string;
-    recepientPID: string;
-  };
-  receiptData: {
-    receiptName: string;
-    date: string;
-    dateOfFiscalization: string;
-    VAT: boolean;
-    note: string;
-    contact: string;
-    contact_name: string;
-    email: string;
-    totalCurrency: string;
-    services: {
-      name: string;
-      price: string;
-      ammount: string;
-      total: string;
-    }[];
-  };
+  apartmentData: TransactionReceiptData["apartmentData"];
+  recepientData: TransactionReceiptData["recepientData"];
+  receiptData: TransactionReceiptData["receiptData"];
 };
 
 const ReceiptTemplate = (props: Props): JSX.Element => {
