+++ src/components/Shared/Receipt/Templates/i18n/TransactionReceipt.en-US.json
@@ -1,7 +1,7 @@
 {
   "date_of_receipt": "Date of receipt",
   "service_name": "Service name",
-  "ammount": "Ammount",
+  "amount": "Amount",
   "price": "Price",
   "total": "Total",
   "total_price": "Total price",
